@@ -1,16 +1,7 @@
 apply plugin: 'java'
 apply plugin: 'application'
-sourceCompatibility = 1.10
-targetCompatibility = 1.10
+sourceCompatibility = 1.11
+targetCompatibility = 1.11
 
 mainClassName = 'midigen.Main'
-
-sourceSets {
-  main {
-    resources {
-        srcDirs = ["src/main/java"]
-        includes = ["**/*.fxml"]
-    }
-  }
-}
 
