@@ -3,19 +3,16 @@ package midigen;
 import javax.sound.midi.*;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.Collectors;
 
 public class Player {
-    private static final List<Integer> majorScale = List.of(0, 2, 4, 5, 7, 9, 11, 12).stream().map(note -> note + 60).collect(Collectors.toList());
+    private static final List<Integer> majorScaleOffsets = List.of(0, 2, 4, 5, 7, 9, 11, 12).stream().map(note -> note + 60).collect(Collectors.toList());
 
     private MidiDevice device = null;
-    private List<Integer> sequence;
+    private List<Tone> sequence;
 
     public Player() {
-        List<Object> synthInfos = new ArrayList<>();
-
         MidiDevice.Info[] infos = MidiSystem.getMidiDeviceInfo();
         for (int i = 0; i < infos.length; i++) {
             try {
@@ -23,13 +20,14 @@ public class Player {
             } catch (MidiUnavailableException e) {
                 e.printStackTrace();
             }
-
             if (device instanceof Synthesizer) {
-                synthInfos.add(infos[i]);
+                break;
             }
         }
 
         if (!(device.isOpen())) {
+            System.out.println("opening: " + device.getDeviceInfo().getDescription());
+
             try {
                 device.open();
             } catch (MidiUnavailableException e) {
@@ -37,17 +35,23 @@ public class Player {
             }
         }
 
-        newSequence();
+        randomSequence();
     }
 
-    public void newSequence() {
-        var newSeq = new ArrayList<Integer>();
+    public void randomSequence() {
+        var newSeq = new ArrayList<Tone>();
 
         for (var i = 0; i < 3; i++) {
-            newSeq.add(majorScale.get(ThreadLocalRandom.current().nextInt(5)));
+            newSeq.add(new Tone(majorScaleOffsets.get(ThreadLocalRandom.current().nextInt(5)), 50));
         }
 
         sequence = newSeq;
+    }
+
+    public void setScale(int i, int time) {
+        sequence = majorScaleOffsets.stream()
+                .map(note -> new Tone(note + i, time))
+                .collect(Collectors.toList());
     }
 
     public void play() {
@@ -57,25 +61,23 @@ public class Player {
         }
     }
 
-    private void play(int frequency) {
+    private void play(Tone tone) {
         ShortMessage myMsg = new ShortMessage();
-        // Start playing the note Middle C (60),
-        // moderately loud (velocity = 93).
         try {
-            myMsg.setMessage(ShortMessage.NOTE_ON, 0, frequency, 93);
+            myMsg.setMessage(ShortMessage.NOTE_ON, 0, tone.getPitch(), 100);
 
             long timeStamp = -1;
-            Receiver rcvr = MidiSystem.getReceiver();
-            rcvr.send(myMsg, timeStamp);
-            Thread.sleep(250);
+            Receiver receiver = MidiSystem.getReceiver();
+            receiver.send(myMsg, timeStamp);
+            Thread.sleep(tone.getTime());
 
-            myMsg.setMessage(ShortMessage.NOTE_OFF, 0, frequency, 93);
-            rcvr.send(myMsg, timeStamp);
-            Thread.sleep(50);
+            myMsg.setMessage(ShortMessage.NOTE_OFF, 0, tone.getPitch(), 100);
+            receiver.send(myMsg, timeStamp);
         } catch (InvalidMidiDataException | InterruptedException e) {
             e.printStackTrace();
         } catch (MidiUnavailableException e) {
             e.printStackTrace();
         }
     }
+
 }
